refactor: PrizeStore를 WinningPrizes로 리팩토링
